# `torch.tensordot`: performance improvements when contracting to a scalar.