--- a/test/distributed/tensor/test_dtensor_ops.py
+++ b/test/distributed/tensor/test_dtensor_ops.py
@@ -131,6 +131,7 @@ dtensor_fails = {
     xfail("cummin"),
     xfail("diagonal_scatter"),
     xfail("dist"),
+    xfail("dot"),
     xfail("empty"),
     xfail("empty_strided"),
     xfail("empty_like"),