@nikitaved 2025-01-30
The distributed issues are real - `aten.dot.default` is missing a sharding strategy. Shall it be added, or what would be the best way to fallback to the sharding-friendly implementation when in the distributed mode?