@bdhirsh 2025-01-30
@albanD who do you think would be a good reviewer for perf improvement PRs like this?