@nikitaved 2025-02-04
Hey, @albanD , thank you for the review! The distributed failures are real. Do you mind if I xfail `tensordot` for now, and then implement a sharding strategy for `dot` in a follow-up?