@albanD 2025-02-04
breaking tensordot on DTensor doesn't sound right no? :p That might be very disruptive.
Is there an easy sharding strategy we can add here? Otherwise we might have to skip this fast path for subclass using at::isTensorSubclassLike(tensor) to check.