@nikitaved 2025-03-12
Alternatively, I could change the dispatcher destination, otherwise I have no the ability to debug MPS (and the issue could be in the dispatcher?).