@albanD 2025-03-12
If the new sample being added was broken on MPS before this change already. It's fine to keep it broken on MPS after this change.
But IIUC, your change should have no impact at all on MPS behavior right?