@ngimel 2025-05-02
The error is in the new path, which will produce 0d tensor no matter what (`return at::dot(t1.flatten(), t2.flatten());`), whereas the sample in question (and many others) wants (1,1), or any number of 1-sized dims really. You need to reshape your output. Side note: the fact that we don't have tests that would actually catch this without MPS is a shame, OpInfos are not all they are cracked up to be. one-off test again is in order.