@ngimel 2025-05-02
Another side note: pytorch docs for tensordot are absolutely useless, I could not parse the sum expression no matter how I tried. Numpy documentation though is better `The shape of the result consists of the non-contracted axes of the first tensor, followed by the non-contracted axes of the second.` https://numpy.org/doc/stable/reference/generated/numpy.tensordot.html, so when contracting 2 3d tensors over 2 dimensions, each of them has 1 non-contracted dimension, so output is 2d