@nikitaved 2025-05-03
Thank you, @ngimel , for the insight. Let us use NumPy then as the source of truth? Yeah, I also got confused by the sum semantics -- one other way to make it easier is to assume that each dimension contraction follows the rules of matrix multiplication. Regarding the consistency and the sample -- do you want it to be fixed here, or could that be done in a follow-up?