@ngimel 2025-05-03
Note, today's behavior is consistent with numpy documentation, and without your PR cuda and cpu produce 2d result, as expected, for this sample. So you need to do correct reshaping in your PR, and I'd prefer that you add tests for this in this PR also, the fact that our existing tests don't catch incorrectly-dimensioned output is pretty bad.