@nikitaved 2025-05-03
The shape issue is resolved, I indeed forgot to properly reshape as per computed result shape. I will add the test for that as well.