@nikitaved 2025-05-08
@ngimel , could you please suggest a strategy for the testing? Do I have to test against a meta-fuction, or just some simple targeted test would do?