@ngimel 2025-05-08
Targeted test will do. Meta function is not there because tensordot is a composite op